Implement a :skip reader
When one wants to discard a couple of bytes (for example, due to
padding), using a dummy field to hold it and ignore it is not
efficient. The :skip reader helps in this case, as it simply moves the
position of the buffer further away, without reading a single byte.

This patch also implements a read-spec change, allowing one to leave
out the key for any skipped fields:

 (read-spec buffer [:header [:magic [:string 4]
                             :skip 16 ; padding
                             :tail-offset 64]])

Of course, writing out a field name also works, but the field will not
be included in the result in this case, either:

 (read-spec buffer [:header [:magic [:string 4]
                             :padding [:skip 16]
                             :tail-offset 64]])

Signed-off-by: Gergely Nagy <algernon@madhouse-project.org>