Fix for proper :skip decoding with decode-blob
With decode-blob, constructs like [:skip 4] is allowed. However, the
skip-or-decode function a few things were delegated to, returned nil
when something was skipped, thereby discarding the map assembled so
far.

This patch adds a test case to guard against that, and also fixes
skip-or-decode to return the unchanged map after skipping bytes.

Signed-off-by: Gergely Nagy <algernon@madhouse-project.org>