Support optional params for the type in decode-blob-array
This patch makes the following possible:

  (decode-blob-array buffer :prefixed :string :uint32)

That is, any arguments after the type, will be passed on to the frame
decoder, allowing for greater flexibility. The main reason for this
feature is that writing one-liner frame decoder methods that
dispatched to another decoder type with a few extra, static arguments,
was becoming painful.

With this feature, these kinds of light and pointless wrappers can be
removed.

Signed-off-by: Gergely Nagy <algernon@madhouse-project.org>